% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality_summary.R
\name{airquality_summary}
\alias{airquality_summary}
\title{airquality_summary}
\usage{
airquality_summary(col_name)
}
\arguments{
\item{col_name}{A character string specifying the column name}
}
\value{
Summary statistics of the column
}
\description{
Summarize a column from the airquality dataset stored in the package
}
\examples{
airquality_summary("Ozone")
airquality_summary("Temp")
}
