% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{plot_box}
\usage{
plot_box(col_name)
}
\arguments{
\item{col_name}{A character string specifying the column to plot}
}
\description{
Create a boxplot of a column from the airquality dataset stored in the package
}
\examples{
plot_box("Ozone")
}
