% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{plot_line}
\usage{
plot_line(col_name)
}
\arguments{
\item{col_name}{The column name to plot}
}
\description{
Create a line plot of a column over time (Month-Day)
}
\examples{
plot_line("Ozone")
}
