% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(x_col, y_col)
}
\arguments{
\item{x_col}{The x-axis column name}

\item{y_col}{The y-axis column name}
}
\description{
Create a scatter plot of two columns from airquality
}
\examples{
plot_scatter("Ozone", "Temp")
}
